[Refactor] 상수화, Enum (Rank, String) 파일 분리
