+++ src/main/kotlin/data/Constant.kt
@@ -1,0 +1,31 @@
+package data
+
+// Lotto
+const val START_NUMBER_ONE = 1
+const val END_NUMBER_FORTY_FIVE = 45
+const val NONE_MATCH = 0
+const val ONE_MATCH = 1
+const val TWO_MATCH = 2
+const val THREE_MATCH = 3
+const val FOUR_MATCH = 4
+const val FIVE_MATCH = 5
+const val SIX_MATCH = 6
+const val FIVE_BONUS_MATCH = 7
+const val TWO = 2
+
+// Purchase
+const val LOTTERY_PRICE = 1000
+
+// Winning
+const val NONE = 0
+const val FIVE_THOUS = 5_000
+const val FIFTY_THOUS = 50_000
+const val ONE_MILLI_FIVE_HUND_THOUS = 1_500_000
+const val THIRTY_MILLI = 30_000_000
+const val TWO_BILLI = 2_000_000_000
+const val PERCENTAGE = 100
+
+// PurchaseTest
+const val TWO_THOUS_FIVE_HUND = 2500
+const val LOTTO_TEN_QUANTITY = 10
+const val TEN_THOUS = 10_000
