+++ src/test/kotlin/lotto/WinningTest.kt
@@ -1,5 +1,6 @@
 package lotto
 
+import data.*
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
